nuit not working

+++ app/src/main/java/com/foxden/fitnessapp/ui/components/NutritionProgress.kt
@@ -24,12 +24,8 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.foxden.fitnessapp.data.NutritionLog
-import com.foxden.fitnessapp.ui.GetGoalsData
-import com.foxden.fitnessapp.ui.theme.DarkBlue
-import java.time.LocalDate
 import com.foxden.fitnessapp.data.SettingsDataStoreManager
+import com.foxden.fitnessapp.ui.GetGoalsData
 
 @RequiresApi(Build.VERSION_CODES.O)
 @OptIn(ExperimentalComposeUiApi::class)
@@ -39,7 +35,7 @@ fun NutritionProgress(numCalories: Int) {
     val context = LocalContext.current
     val dataStoreManager = SettingsDataStoreManager(context)
 
-    var calorieTarget by remember { mutableStateOf(0.0f) }
+    var calorieTarget by remember { mutableStateOf(0) }
 
     LaunchedEffect(Unit) {
         GetGoalsData(
@@ -54,7 +50,7 @@ fun NutritionProgress(numCalories: Int) {
         )
     }
 
-    val progress = (numCalories / calorieTarget) //: Float = calorieTarget / numCalories
+    val progress = (numCalories / calorieTarget).toFloat() //: Float = calorieTarget / numCalories
 
 
     val containerModifier = Modifier
